Module Ocaml_version.Has

Test whether a release has a given feature.

val bytes : t -> bool

bytes t will return true if that release has a bytes type. Note that opam provides a bytes compatibility package for older releases.

val arch : arch -> t -> bool

arch a t will return true if architecture a is supported on release t.