Module Opam.V2

Opam 2.0 functions

val name : t -> string

package t returns the opam2 package for that compiler version.

val variant_switch : t -> Configure_options.o list -> t

variant_switch t cs returns an OCaml version t whose variant version field reflects the configuration options in cs

val switches : arch -> t -> t list

switches arch t returns the list of opam switches that are available for that compiler and architecture combination. For example, an x86_64 one would yield a list that includes flambda and the flambda-with-frame-pointer combinations.