Module Ocaml_version.Releases

Values representing official releases of OCaml.

val v4_00_1 : t

Version 4.00.1

val v4_00 : t

Latest release in the 4.00.x series

val v4_01_0 : t

Version 4.01.0

val v4_01 : t

Latest release in the 4.01.x series

val v4_02_0 : t

Version 4.02.0

val v4_02_1 : t

Version 4.02.1

val v4_02_2 : t

Version 4.02.2

val v4_02_3 : t

Version 4.02.3

val v4_02 : t

Latest release in the 4.02.x series

val v4_03_0 : t

Version 4.03.0

val v4_03 : t

Latest release in the 4.03.x series

val v4_04_0 : t

Version 4.04.0

val v4_04_1 : t

Version 4.04.1

val v4_04_2 : t

Version 4.04.2

val v4_04 : t

Latest release in the 4.04.x series

val v4_05_0 : t

Version 4.05.0

val v4_05 : t

Latest release in the 4.05.x series

val v4_06_0 : t

Version 4.06.0

val v4_06_1 : t

Version 4.06.1

val v4_06 : t

Latest release in the 4.06.x series

val v4_07_0 : t

Version 4.07.0

val v4_07_1 : t

Version 4.07.1

val v4_07 : t

Latest release in the 4.07.x series

val v4_08_0 : t

Versior 4.08.0

val v4_08 : t

Latest release in the 4.08.x series

val v4_09_0 : t

Version 4.09.0

val v4_09_1 : t

Version 4.09.1

val v4_09 : t

Latest release in the 4.09.x series

val v4_10_0 : t

Version 4.10.0

val v4_10_1 : t

Version 4.10.1

val v4_10 : t

Latest release in the 4.10.x series

val v4_11_0 : t

Version 4.11.0

val v4_11 : t

Latest release in the 4.11.x series

val v4_12 : t

Latest release in the 4.12.x series

val all_patches : t list

all_patches is an enumeration of all OCaml releases, including every patch release. To get the major and minor releases with the latest patch version, use all instead.

val all : t list

all is an enumeration of all the OCaml releases, with the latest patch versions in each major and minor release.

val unreleased_betas : t list

Enumerates the latest beta / release-candidate versions for each unreleased minor OCaml series.

val dev : t list

Enumeration of the latest development OCaml releases. This is usually just one, but may include two active dev versions if a release branch has just been cut.

val latest : t

latest is the most recent stable release of OCaml.

val recent : t list

recent is the last eight releases of OCaml, with each at the latest patch level. This is the set that is most reliably tested in the opam package repository.

val recent_with_dev : t list

recent_with_dev are the last eight stable releases of OCaml and the latest development branches.

val is_dev : t -> bool

is_dev t will return true if the release t represents a development release instead of a stable archive.