Module Ocaml_version.Since

Determine which release a feature or architecture first appeared in.

val bytes : t

bytes is the release that the Bytes module first appeared in.

val arch : arch -> t

arch a will return the first release of OCaml that the architecture was reasonably stably supported on.