Module Ocaml_version.Sources

Values relating to the source code and version control of OCaml

val trunk : t

trunk is the version of the development head of the OCaml git repository.

val git_tag : t -> string